+++ src/lib/shim/shim.c
@@ -32,9 +32,6 @@
 #include "lib/shim/shim_tls.h"
 #include "main/host/syscall_numbers.h" // for SYS_shadow_* defs
 
-// Whether Shadow is using the shim-side syscall handler optimization.
-static bool _using_shim_syscall_handler = true;
-
 // This thread's IPC block, for communication with Shadow.
 static ShMemBlock* _shim_ipcDataBlk() {
     static ShimTlsVar v = {0};
@@ -126,23 +123,12 @@ bool shim_interpositionEnabled() {
     return !*_shim_allowNativeSyscallsFlag();
 }
 
-bool shim_use_syscall_handler() { return _using_shim_syscall_handler; }
-
 static bool _running_in_shadow = false;
 
 // Whether we're running in Shadow. When this is false the shim mostly
 // does nothing. This can be useful e.g. for programs that are Shadow-aware
 // and link against the shim so that they can call Shadow APIs.
 static void _shim_setRunningInShadow() { _running_in_shadow = getenv("SHADOW_SPAWNED") != NULL; }
-
-static void _set_use_shim_syscall_handler() {
-    const char* shim_syscall_str = getenv("SHADOW_DISABLE_SHIM_SYSCALL");
-    if (shim_syscall_str && !strcmp(shim_syscall_str, "TRUE")) {
-        _using_shim_syscall_handler = false;
-    } else {
-        _using_shim_syscall_handler = true;
-    }
-}
 
 static void** _shim_signal_stack() {
     static ShimTlsVar stack_var = {0};
@@ -487,7 +473,6 @@ __attribute__((constructor)) void _shim_load() {
         logger_setLevel(logger_getDefault(), LOGLEVEL_WARNING);
 
         _shim_setRunningInShadow();
-        _set_use_shim_syscall_handler();
     }
 
     if (!_running_in_shadow) {