+++ src/main/host/host.rs
@@ -74,7 +74,6 @@ pub struct HostParameters {
     pub unblocked_syscall_latency: SimulationTime,
     pub unblocked_vdso_latency: SimulationTime,
     pub use_legacy_working_dir: bool,
-    pub use_shim_syscall_handler: bool,
     pub strace_logging_options: Option<FmtOptions>,
 }
 
@@ -388,7 +387,6 @@ impl Host {
             argv,
             pause_for_debugging,
             self.params.use_legacy_working_dir,
-            self.params.use_shim_syscall_handler,
             self.params.strace_logging_options,
         );